notifications

+++ calls/deletepost.php
@@ -1,9 +1,9 @@
 <?php
-
+session_start();
 require_once('../global.php');
 
 $database = new Database;
-if (isset($_GET['idpost'])) {
+if (Tools::isLogged() && isset($_GET['idpost'])) {
 	if ($database->deletePost($_GET['idpost'])) {
 		echo "ok";
 	}
